refactor(core): extract wave1 crypto bridge and data-layer modules (#6231)
