+++ crates/kamn-bridges/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "kamn-bridges"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[lib]
+path = "src/lib.rs"
+
+[dependencies]
+
+[lints]
+workspace = true
