+++ crates/kamn-bridges/src/lib.rs
@@ -1,0 +1,5 @@
+#![warn(missing_docs)]
+//! Extracted bridge-domain surfaces from `kamn-core`.
+
+/// Cross-chain receipt proof normalization and finality mapping contracts.
+pub mod cross_chain_receipt;
