+++ crates/kamn-core/Cargo.toml
@@ -13,6 +13,9 @@ live-https = ["dep:rustls", "dep:rustls-pemfile", "dep:webpki-roots"]
 libp2p-live-transport = ["dep:libp2p", "dep:tokio"]
 
 [dependencies]
+kamn-bridges = { path = "../kamn-bridges" }
+kamn-crypto = { path = "../kamn-crypto" }
+kamn-data-layer = { path = "../kamn-data-layer" }
 kamn-kolme = { path = "../kamn-kolme" }
 kamn-live-probe-matrix = { path = "../kamn-live-probe-matrix" }
 kamn-runtime-guards = { path = "../kamn-runtime-guards" }
