+++ crates/kamn-crypto/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "kamn-crypto"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[lib]
+path = "src/lib.rs"
+
+[dependencies]
+chacha20poly1305 = "0.10.1"
+sha2 = "0.10.8"
+x25519-dalek = { version = "2.0.1", features = ["static_secrets"] }
+
+[lints]
+workspace = true
