+++ crates/kamn-crypto/src/lib.rs
@@ -1,0 +1,5 @@
+#![warn(missing_docs)]
+//! Extracted cryptography surfaces from `kamn-core`.
+
+/// Direct-message encryption/decryption contracts and error semantics.
+pub mod direct_message_crypto;
