+++ crates/kamn-data-layer/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "kamn-data-layer"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[lib]
+path = "src/lib.rs"
+
+[dependencies]
+k256 = { version = "0.13.4", features = ["ecdsa"] }
+
+[lints]
+workspace = true
