+++ crates/kamn-data-layer/src/lib.rs
@@ -1,0 +1,5 @@
+#![warn(missing_docs)]
+//! Extracted data-layer shared helpers from `kamn-core`.
+
+/// Shared SHA-256 helpers used by data-layer hash-chain contracts.
+pub mod data_layer_hashing;
